requirements.txt
# Libraries needed to run this project
tensorflow>=2.8
numpy
streamlit>=1.10
Pillow
matplotlib
